pyproject.toml
[project]
name = "career-guidance"
version = "0.1.0"
description = "Tool to guide job seekers supported by AI"
authors = [
    {name = "Pradeepaa001",email = "mpradeepaa2020@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai",
    "python-dotenv",
    "spacy",
    "pymupdf",
    "python-docx",
    "PyPDF2",
    "pyaudio",
    "numpy",
    "whisper",
"SpeechRecognition",
"pyttsx3",
"python-docx",
"fastapi",
"uvicorn",
 "python-multipart"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
